fix(test): split lesson-anchor-coherence.test.ts to hold Section 4 razor
The coherence test file was 523 lines after this PR added the
essay-list mount class (FX598 extension) + FX602 glossary-class
validation + FX612 SWE-vocab dominance check on top of the v1
baseline. The #390 seal claimed "every new file ≤ 250" but this
v1-grandfathered file the seal materially extended was over.

Split into three sibling files (all ≤ 250):
- lesson-anchor-coherence-helpers.ts (234) — shared model +
  source-scan helpers; no .test.ts suffix so mocha does not load it
- lesson-anchor-coherence.test.ts (153) — FX598 + FX602 + essay-list
  + UI_MANIFEST presence (13 assertions preserved verbatim)
- lesson-anchor-vocab.test.ts (66) — FX612 SWE-vocab dominance
  (2 assertions preserved verbatim; vocab logic is independent of
  the source-scan model so the split is clean)

No semantic change. All 13 prior assertions pass. vscode-test still
2866 passing / 1 pending / 0 failing.

META_LEDGER entry #391 records the correction (entry #390 stays
intact per the chain-integrity rule). SYSTEM_STATE razor claim and
FEATURE_INDEX FX598/FX602/FX612 rows updated to cite the new paths.

Caught by post-PR-submission Codex review.
