+++ docs/SYSTEM_STATE.md
@@ -18,7 +18,7 @@ The v5.2.0 release-gating Learn-tab surface, rebuilt from the operator's reframe
 - **Compliance enforcement** — extended FX598 coherence check with essay-list mount class; new FX612 SWE-craft vocabulary dominance check (≥ 3× FailSafe-governance vocabulary count in `learn.essay.*` bodies; hostile fixture caught).
 - **Scrap (rejected v1 + v3)** — `devCycleTrack.ts`, `guided-dev-cycle.js`, the FX603/604/605/607 tests, `docs/GUIDED_DEV_CYCLE.md` all deleted. `bundle.cjs` / `copy-ui-js.cjs` browser-ESM emit reverted from `devCycleTrack` to `lessonTriggers`. The 4 v1 governance-moment lesson literals (`governance-mode`, `shield.plan/audit/substantiate`) carry forward — they power the v1 Settings governance-mode card + Monitor SHIELD phase tracker, NOT the Learn tab.
 
-Verification: `tsc` clean; lint 0 errors (114 warnings, +1 vs v3 baseline); `vscode-test` **2866 passing / 1 pending / 0 failing** (post-review remediation added 28 new assertions across 3 cuts: table-driven HIGH_BLAST_RADIUS fixtures, cumulative-cap tests, passive-state scope test, curriculum-always-renders test, cap-persistence integration test, Mark-as-read semantic, disabled-Learn cap-budget preservation, and hidden-tab cap-budget preservation); Playwright FX611 (new) + FX601 (modified) **4/4**. Section 4 Razor: every new file ≤ 250 (`lessonTriggers.ts` 250 at-cap; `lessonTriggers.test.ts` 176 + sibling `lessonTriggers-caps.test.ts` 90 — applyCaps suite split post-Cut-2; `learn.js` 173; `learn-essay-list.js` 187; `learn-tab.test.ts` 61 + sibling `learn-tab-triggers.test.ts` 175 + shared helper `learn-tab-helpers.ts` 49 — trigger-integration suite split post-Cut-3; `learn-essay-list.test.ts` 142). Audit: cycle-1 VETO (2 findings: enumDescriptions infrastructure-mismatch + FX613 lacked automated test) → both remediated → cycle-2 PASS (12/12) via independent architect-reviewer subagent. Compliance: fresh determination — stays outside EU AI Act Annex III; more strongly aligned with Article 4; GDPR client-side-only session timing binding; 2 binding escalation triggers (scoring/inference → re-ideate; curriculum-in-extension → escalate to Pro). `package.json` remains `5.1.8`; v5.2.0 ships at `/qor-repo-release`.
+Verification: `tsc` clean; lint 0 errors (114 warnings, +1 vs v3 baseline); `vscode-test` **2866 passing / 1 pending / 0 failing** (post-review remediation added 28 new assertions across 3 cuts: table-driven HIGH_BLAST_RADIUS fixtures, cumulative-cap tests, passive-state scope test, curriculum-always-renders test, cap-persistence integration test, Mark-as-read semantic, disabled-Learn cap-budget preservation, and hidden-tab cap-budget preservation); Playwright FX611 (new) + FX601 (modified) **4/4**. Section 4 Razor: every new/modified Learn-v2 file ≤ 250 (`lessonTriggers.ts` 250 at-cap; `lessonTriggers.test.ts` 176 + sibling `lessonTriggers-caps.test.ts` 90 — applyCaps suite split post-Cut-2; `learn.js` 173; `learn-essay-list.js` 187; `learn-tab.test.ts` 61 + sibling `learn-tab-triggers.test.ts` 175 + shared helper `learn-tab-helpers.ts` 49 — trigger-integration suite split post-Cut-3; `learn-essay-list.test.ts` 142; `lesson-anchor-coherence.test.ts` 153 + sibling `lesson-anchor-vocab.test.ts` 66 + shared helper `lesson-anchor-coherence-helpers.ts` 234 — FX612 vocab suite + model/scan helpers split out post-Cut-5 after the file passed 523 lines once the essay-list mount class + SWE-vocab dominance check were folded in; the original was a v1-grandfathered file that grew past the razor when this PR materially extended it). Audit: cycle-1 VETO (2 findings: enumDescriptions infrastructure-mismatch + FX613 lacked automated test) → both remediated → cycle-2 PASS (12/12) via independent architect-reviewer subagent. Compliance: fresh determination — stays outside EU AI Act Annex III; more strongly aligned with Article 4; GDPR client-side-only session timing binding; 2 binding escalation triggers (scoring/inference → re-ideate; curriculum-in-extension → escalate to Pro). `package.json` remains `5.1.8`; v5.2.0 ships at `/qor-repo-release`.
 
 Implementation-time deviations (autonomous, logged): (1) plan v4 Phase 1 said "REPLACE" the 4 governance-moment literals but the §Anchor table marked them KEEP — three evidence signals (plan §Anchor, research-brief Finding 4a, functional dependency) agreed; autonomous fix: ADD the 5 SWE essays alongside the 4 v1 carry-forwards. (2) Plan v4 Phase 4 didn't specify the essay-list mount class in the coherence test — added parallel to the existing glossary class extension.
